Move presentationMode eventHandling to separate folder and class
